refact: (post-review) Move class_path hash to BrokerResolver class
