+++ lib/pipefy_message.rb
@@ -17,18 +17,4 @@ require "active_support/core_ext/string/inflections"
 # PipefyMessage abstraction async process
 ##
 module PipefyMessage
-  def self.class_path
-    {
-      aws: {
-        publisher: {
-          class_name: "PipefyMessage::Providers::AwsClient::SnsBroker",
-          relative_path: "providers/aws_client/sns_broker"
-        },
-        consumer: {
-          class_name: "PipefyMessage::Providers::AwsClient::SqsBroker",
-          relative_path: "providers/aws_client/sqs_broker"
-        }
-      }
-    }
-  end
 end
